Make several semantic and style changes to SrvServiceDiscovery + test
 * Rename SrvResolver and interface to use standard semantics
 * Add visual line separation between arrange, act, assert test blocks
