+++ app/autodiscovery/src/main/java/com/fsck/k9/autodiscovery/KoinModule.kt
@@ -2,13 +2,13 @@ package com.fsck.k9.autodiscovery
 
 import com.fsck.k9.autodiscovery.providersxml.ProvidersXmlDiscovery
 import com.fsck.k9.autodiscovery.providersxml.ProvidersXmlProvider
-import com.fsck.k9.autodiscovery.srvrecords.SrvResolver
+import com.fsck.k9.autodiscovery.srvrecords.MiniDnsSrvResolver
 import com.fsck.k9.autodiscovery.srvrecords.SrvServiceDiscovery
 import org.koin.dsl.module
 
 val autodiscoveryModule = module {
     factory { ProvidersXmlProvider(get()) }
     factory { ProvidersXmlDiscovery(get(), get()) }
-    single { SrvResolver() }
+    single { MiniDnsSrvResolver() }
     factory { SrvServiceDiscovery(get()) }
 }
